website/requirements.txt
attrs==21.2.0
bcrypt==3.1.7
cffi==1.15.0
click==8.0.3
coverage==6.2
csvalidate==1.1.1
Flask==2.0.2
Flask-CSV==1.2.0
Flask-Login==0.4.1
Flask-Mako==0.4
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==1.2.2
Werkzeug==2.0.2
